large restructure and remove deprecated
