+++ yucca/evaluation/obj_metrics.py
@@ -1,6 +1,6 @@
 import numpy as np
 import cc3d
-from yucca.utils.nib_utils import get_nib_spacing
+from yucca.functional.utils.nib_utils import get_nib_spacing
 from yucca.evaluation.metrics import sensitivity, precision, f1
