+++ yucca/evaluation/surface_metrics.py
@@ -1,5 +1,5 @@
 import numpy as np
-from yucca.utils.nib_utils import get_nib_spacing
+from yucca.functional.utils.nib_utils import get_nib_spacing
 from yucca.evaluation.deepmind_surface_distance import metrics
