+++ yucca/functional/array_operations/bounding_boxes.py
