+++ yucca/functional/array_operations/cropping_and_padding.py
