+++ yucca/functional/array_operations/transpose.py
