+++ yucca/network_architectures/networks/YuccaNet.py
@@ -1,6 +1,6 @@
 import torch
 from torch import nn
-from yucca.utils.torch_utils import get_available_device
+from yucca.functional.utils.torch_utils import get_available_device
 from yucca.network_architectures.utils.get_steps_for_sliding_window import (
     get_steps_for_sliding_window,
 )
