+++ yucca/run/run_preprocessing.py
@@ -1,7 +1,7 @@
 import argparse
 import yucca
-from yucca.utils.task_ids import maybe_get_task_from_task_id
-from yucca.utils.files_and_folders import recursive_find_python_class
+from yucca.task_conversion.utils import maybe_get_task_from_task_id
+from yucca.functional.utils.files_and_folders import recursive_find_python_class
 from batchgenerators.utilities.file_and_folder_operations import join
