+++ yucca/task_conversion/utils.py
@@ -1,11 +1,11 @@
 import numpy as np
 import os
 import shutil
+import nibabel as nib
 from yucca.paths import yucca_raw_data
 from typing import Literal
-from batchgenerators.utilities.file_and_folder_operations import save_json, subfiles, join
+from batchgenerators.utilities.file_and_folder_operations import save_json, subfiles, join, subdirs
 from tqdm import tqdm
-import nibabel as nib
 from pathlib import Path
 
 
@@ -115,3 +115,24 @@ def generate_dataset_json(
             "Proceeding anyways..."
         )
     save_json(json_dict, os.path.join(output_file))
+
+
+def maybe_get_task_from_task_id(task_id: str | int):
+    task_id = str(task_id)
+    tasks = subdirs(yucca_raw_data, join=False)
+
+    # Check if name is already complete
+    if task_id in tasks:
+        return task_id
+
+    # If not, we try to recreate the name
+    # We use the raw_data folder as reference
+    for task in tasks:
+        if task_id.lower() in task.lower():
+            return task
+
+    # If we can't find anything we just return the original, on the offchance that the task does not exist in Raw Data while existing in e.g. Preprocessed
+    print(
+        f"Couldn't find a task called: {task_id} in the raw data folder: {yucca_raw_data}. If your task only exists in e.g. the Preprocessed folder things might still work."
+    )
+    return task_id
